Finish refactor
