+++ dockstore-client/src/main/java/io/dockstore/client/cli/nested/AbstractEntryClient.java
@@ -32,6 +32,7 @@ import io.cwl.avro.CommandLineTool;
 import io.cwl.avro.Workflow;
 import io.dockstore.client.Bridge;
 import io.dockstore.client.cli.Client;
+import io.dockstore.client.cwlrunner.CWLRunnerFactory;
 import io.dockstore.common.FileProvisioning;
 import io.dockstore.common.Utilities;
 import io.dockstore.common.WDLFileProvisioning;
@@ -114,7 +115,6 @@ import static io.dockstore.client.cli.Client.SCRIPT;
 public abstract class AbstractEntryClient {
     private static final String CROMWELL_LOCATION = "https://github.com/broadinstitute/cromwell/releases/download/0.21/cromwell-0.21.jar";
     private static final Logger LOG = LoggerFactory.getLogger(AbstractEntryClient.class);
-    private static final CWL CWL_UTIL = new CWL();
     boolean isAdmin = false;
 
     static String getCleanedDescription(String description) {
@@ -126,6 +126,11 @@ public abstract class AbstractEntryClient {
             }
         }
         return description;
+    }
+
+    private CWL getCwlUtil() {
+        String cwlrunner = CWLRunnerFactory.getCWLRunner();
+        return new CWL(cwlrunner.equalsIgnoreCase(CWLRunnerFactory.CWLRunner.BUNNY.toString()));
     }
 
     public abstract String getConfigFile();
@@ -558,7 +563,7 @@ public abstract class AbstractEntryClient {
         } else {
             final String cwlPath = reqVal(args, "--cwl");
           
-            final ImmutablePair<String, String> output = CWL_UTIL.parseCWL(cwlPath);
+            final ImmutablePair<String, String> output = getCwlUtil().parseCWL(cwlPath);
           
             // do not continue to convert to json if cwl is invalid
             if (!validateCWL(cwlPath)) {
@@ -566,7 +571,7 @@ public abstract class AbstractEntryClient {
             }
 
             try {
-                final Map<String, Object> runJson = CWL_UTIL.extractRunJson(output.getLeft());
+                final Map<String, Object> runJson = getCwlUtil().extractRunJson(output.getLeft());
                 if (json) {
                     final Gson gson = io.cwl.avro.CWL.getTypeSafeCWLToolDocument();
                     out(gson.toJson(runJson));
@@ -1413,8 +1418,8 @@ public abstract class AbstractEntryClient {
 
         if (descriptor.equals(CWL_STRING)) {
             // need to suppress output
-            final ImmutablePair<String, String> output = CWL_UTIL.parseCWL(tempDescriptor.getAbsolutePath());
-            final Map<String, Object> stringObjectMap = CWL_UTIL.extractRunJson(output.getLeft());
+            final ImmutablePair<String, String> output = getCwlUtil().parseCWL(tempDescriptor.getAbsolutePath());
+            final Map<String, Object> stringObjectMap = getCwlUtil().extractRunJson(output.getLeft());
             if (json) {
                 try {
                     final Gson gson = CWL.getTypeSafeCWLToolDocument();
@@ -1426,7 +1431,7 @@ public abstract class AbstractEntryClient {
                 }
             } else {
                 // re-arrange as rows and columns
-                final Map<String, String> typeMap = CWL_UTIL.extractCWLTypes(output.getLeft());
+                final Map<String, String> typeMap = getCwlUtil().extractCWLTypes(output.getLeft());
                 final List<String> headers = new ArrayList<>();
                 final List<String> types = new ArrayList<>();
                 final List<String> entries = new ArrayList<>();
