+++ dockstore-client/src/main/java/io/github/collaboratory/LauncherCWL.java
@@ -49,6 +49,7 @@ import io.cwl.avro.CommandOutputParameter;
 import io.cwl.avro.Workflow;
 import io.cwl.avro.WorkflowOutputParameter;
 import io.dockstore.client.cwlrunner.CWLRunnerFactory;
+import io.dockstore.client.cwlrunner.CWLRunnerInterface;
 import io.dockstore.common.FileProvisioning;
 import io.dockstore.common.Utilities;
 import org.apache.commons.cli.CommandLine;
@@ -131,7 +132,9 @@ public class LauncherCWL {
         config = Utilities.parseConfig(configFilePath);
 
         // parse the CWL tool definition without validation
-        CWL cwlUtil = new CWL();
+        CWLRunnerFactory.setConfig(config);
+        String cwlRunner = CWLRunnerFactory.getCWLRunner();
+        CWL cwlUtil = new CWL(cwlRunner.equalsIgnoreCase(CWLRunnerFactory.CWLRunner.BUNNY.toString()));
         final String imageDescriptorContent = cwlUtil.parseCWL(imageDescriptorPath).getLeft();
         Object cwlObject;
         try {
@@ -429,7 +432,10 @@ public class LauncherCWL {
         try {
             Files.createDirectories(Paths.get(workingDir));
             try {
-                Utilities.executeCommand("setfacl -d -m o::rwx " + workingDir);
+                boolean useBunny = config.getString("cwlrunner", "cwltool").equalsIgnoreCase(CWLRunnerFactory.CWLRunner.BUNNY.toString());
+                if (useBunny) {
+                    Utilities.executeCommand("setfacl -d -m o::rwx " + workingDir);
+                }
             } catch (Exception e) {
                 System.err.println("WARNING: Unable to set default permissions on working dir, may "
                         + "result in problems with Docker containers that change users : setfacl -d -m o::rwx " + workingDir);
