+++ dockstore-client/src/test/resources/launcher.cwltool.ini
