+++ pom.xml
@@ -45,7 +45,7 @@
         <aws.version>1.11.86</aws.version>
         <powermock.version>1.6.5</powermock.version>
         <mockito.version>1.10.19</mockito.version>
-        <cwlavro.version>1.0.9-SNAPSHOT</cwlavro.version>
+        <cwlavro.version>2.0.0</cwlavro.version>
         <okhttp.version>3.2.0</okhttp.version>
         <slf4j.version>1.7.21</slf4j.version>
         <swagger-ui.version>2.1.4</swagger-ui.version>
@@ -868,8 +868,6 @@
                                     <failWhenParentIsSnapshot>false</failWhenParentIsSnapshot>
                                     <excludes>
                                         <exclude>io.dockstore:*</exclude>
-                                        <excluse>io.cwl:cwlavro-tools</excluse>
-                                        <excluse>io.cwl:cwlavro-generated</excluse>
                                     </excludes>
                                 </requireReleaseDeps>
                                 <requireUpperBoundDeps />
